+++ src/installer/install_tree.rs
@@ -289,8 +289,7 @@ impl<'a> InstallTreeBuilder<'a> {
     /// Tries to find a prebuild for the given package.
     /// Returns true if a prebuild is found, false otherwise.
     fn find_prebuild(&mut self, install_meta: &InstallMeta, package_id: &PackageId) -> Result<bool> {
-        let prebuilds_list =
-            self.repository_manager.read_prebuilds_list(&install_meta.repository_id, &package_id.name, &package_id.version)?;
+        let prebuilds_list = self.repository_manager.read_prebuilds_list(&install_meta.repository_id, &package_id)?;
 
         let Some((prebuild_id, _)) = prebuilds_list.get_best_prebuild(&Target::current()) else {
             return Ok(false);