+++ src/repositories/metadata/provider.rs
@@ -1,12 +1,13 @@
 // SPDX-License-Identifier: GPL-3.0-only
-use std::{hash::Hash, sync::Arc};
+use std::sync::Arc;
 
 use bytes::Bytes;
-use moka::{Equivalent, sync::Cache};
+use moka::sync::Cache;
+use once_cell::sync::OnceCell;
 
 use crate::{
     config::Repository,
-    installer::types::{PackageName, Version},
+    installer::types::{PackageId, PackageName, Version},
     repositories::{
         error::Result,
         metadata::{
@@ -57,12 +58,12 @@ impl MetadataProvider {
 
     /// Reads repository metadata from the repository, containing information about the repository.
     pub fn read_repository_metadata(&self) -> Result<RepositoryMeta> {
-        self.inner.read_repository_metadata()
+        self.cache.get_repository_metadata(|| self.inner.read_repository_metadata())
     }
 
     /// Reads index metadata from the repository, containing all supported packages.
     pub fn read_index_metadata(&self) -> Result<IndexMeta> {
-        self.inner.read_index_metadata()
+        self.cache.get_index_metadata(|| self.inner.read_index_metadata())
     }
 
     /// Reads package metadata from the repository, containing information about the package.
@@ -71,14 +72,16 @@ impl MetadataProvider {
     }
 
     /// Reads the metadata of a certain version of a package, containing dependencies and targets.
-    pub fn read_package_version(&self, package: &PackageName, version: &Version) -> Result<PackageVersionMeta> {
-        self.cache.get_package_version(package, version, || self.inner.read_package_version(package, version))
+    pub fn read_package_version(&self, package_id: &PackageId) -> Result<PackageVersionMeta> {
+        self.cache.get_package_version(package_id, || {
+            self.inner.read_package_version(&package_id.name, &package_id.version)
+        })
     }
 
     /// Reads the list of prebuilds that can be generated for the given version of the package.
     /// Returns `None` if the prebuilds list does not exist.
-    pub fn read_prebuilds_list(&self, package: &PackageName, version: &Version) -> Result<Option<PrebuildsList>> {
-        self.cache.get_prebuilds_list(package, version, || self.inner.read_prebuilds_list(package, version))
+    pub fn read_prebuilds_list(&self, package_id: &PackageId) -> Result<Option<PrebuildsList>> {
+        self.cache.get_prebuilds_list(package_id, || self.inner.read_prebuilds_list(&package_id.name, &package_id.version))
     }
 
     /// Reads the requested file from the repository as bytes.
@@ -97,21 +100,44 @@ fn boxed(provider: Option<impl MetadataProviderImpl + 'static>) -> Option<Box<dy
     provider.map(|provider| Box::new(provider) as Box<dyn MetadataProviderImpl>)
 }
 
+/// Represents the cache of the metadata for a single `MetadataProvider`.
 struct MetadataCache {
+    repository_metadata_cache: OnceCell<RepositoryMeta>,
+    index_metadata_cache: OnceCell<IndexMeta>,
     package_cache: Cache<PackageName, PackageMeta>,
-    package_version_cache: Cache<PackageVersionKey, PackageVersionMeta>,
-    prebuilds_list_cache: Cache<PackageVersionKey, Option<PrebuildsList>>,
+    package_version_cache: Cache<PackageId, PackageVersionMeta>,
+    prebuilds_list_cache: Cache<PackageId, Option<PrebuildsList>>,
 }
 
 impl MetadataCache {
+    /// Creates a new `MetadataCache`.
     pub fn new() -> Self {
         Self {
+            repository_metadata_cache: OnceCell::new(),
+            index_metadata_cache: OnceCell::new(),
             package_cache: Cache::new(100),
             package_version_cache: Cache::new(100),
             prebuilds_list_cache: Cache::new(100),
         }
     }
 
+    /// Gets the `RepositoryMeta` from the cache, or initializes it using the init function.
+    pub fn get_repository_metadata<F>(&self, init: F) -> Result<RepositoryMeta>
+    where
+        F: FnOnce() -> Result<RepositoryMeta>,
+    {
+        self.repository_metadata_cache.get_or_try_init(init).map(Clone::clone)
+    }
+
+    /// Gets the `IndexMeta` from the cache, or initializes it using the init function.
+    pub fn get_index_metadata<F>(&self, init: F) -> Result<IndexMeta>
+    where
+        F: FnOnce() -> Result<IndexMeta>,
+    {
+        self.index_metadata_cache.get_or_try_init(init).map(Clone::clone)
+    }
+
+    /// Gets the `PackageMeta` from the cache, or initializes it using the init function.
     pub fn get_package<F>(&self, package: &PackageName, init: F) -> Result<PackageMeta>
     where
         F: FnOnce() -> Result<PackageMeta>,
@@ -119,51 +145,19 @@ impl MetadataCache {
         self.package_cache.try_get_with_by_ref(package, init).map_err(Arc::unwrap_or_clone)
     }
 
-    pub fn get_package_version<F>(&self, package: &PackageName, version: &Version, init: F) -> Result<PackageVersionMeta>
+    /// Gets the `PackageVersionMeta` from the cache, or initializes it using the init function.
+    pub fn get_package_version<F>(&self, package_id: &PackageId, init: F) -> Result<PackageVersionMeta>
     where
         F: FnOnce() -> Result<PackageVersionMeta>,
     {
-        // Try to get value without cloning key
-        let key = PackageVersionKeyRef(package, version);
-        if let Some(value) = self.package_version_cache.get(&key) {
-            return Ok(value);
-        }
-
-        // Initialize or get value with cloned key, to avoid concurrent fetch calls
-        let key = PackageVersionKey(package.clone(), version.clone());
-        self.package_version_cache.try_get_with(key, init).map_err(Arc::unwrap_or_clone)
+        self.package_version_cache.try_get_with_by_ref(package_id, init).map_err(Arc::unwrap_or_clone)
     }
 
-    pub fn get_prebuilds_list<F>(&self, package: &PackageName, version: &Version, init: F) -> Result<Option<PrebuildsList>>
+    /// Gets the `PrebuildsList` from the cache, or initializes it using the init function.
+    pub fn get_prebuilds_list<F>(&self, package_id: &PackageId, init: F) -> Result<Option<PrebuildsList>>
     where
         F: FnOnce() -> Result<Option<PrebuildsList>>,
     {
-        // Try to get value without cloning key
-        let key = PackageVersionKeyRef(package, version);
-        if let Some(value) = self.prebuilds_list_cache.get(&key) {
-            return Ok(value);
-        }
-
-        // Initialize or get value with cloned key, to avoid concurrent fetch calls
-        let key = PackageVersionKey(package.clone(), version.clone());
-        self.prebuilds_list_cache.try_get_with(key, init).map_err(Arc::unwrap_or_clone)
-    }
-}
-
-#[derive(PartialEq, Eq)]
-struct PackageVersionKey(PackageName, Version);
-
-#[derive(Hash)]
-struct PackageVersionKeyRef<'a>(&'a PackageName, &'a Version);
-
-impl Equivalent<PackageVersionKey> for PackageVersionKeyRef<'_> {
-    fn equivalent(&self, key: &PackageVersionKey) -> bool {
-        self.0 == &key.0 && self.1 == &key.1
-    }
-}
-
-impl Hash for PackageVersionKey {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        PackageVersionKeyRef(&self.0, &self.1).hash(state);
+        self.prebuilds_list_cache.try_get_with_by_ref(package_id, init).map_err(Arc::unwrap_or_clone)
     }
 }